+++ libs/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @crankscript/types
 
+## 2.2.1
+
+### Patch Changes
+
+- Use const enums
+
 ## 2.2.0
 
 ### Minor Changes